Add REST API base
In order for an athlete to log an activity, there will be a front-end for the athlete to enter data. This means there should be a back-end to send the data to, and to store it.

This commit ensures we set up a working REST API as a base for the endpoints.