Add Activities REST Endpoint to Enable Adding Activities
In order for an athlete to log an activity, and that the athlete achieves this through a front-end application, there needs to be an API endpoint to achieve this.

This commit ensures we set up an API endpoint for Activities. The `add` operation is included so that we can add Activities according to the Athlete.

Also included: setting the DB as a global variable, to perform DB operations across the app. (recognised as a hacky solution)