Add Basic Error Handling for Adding an Activity Endpoint
If an athlete enters an incomplete form, or enters the wrong information, then the API should return an appropriate response code and error.

This commit ensures we catch any errors when trying to add the activity. Ideally, we should pass the data through a validator before attempting to add the Activity, however, this can be added in later.

DEBUG mode set to False, otherwise assertion errors occur. This is common issue with Flask.