Add Activity Repository
Part of refactoring the logic out of the Activity API endpoint, as that logic should not live there.

This commit ensures we set up a repository for functions relating the Activity entity. Ideally, we'd want a base repository set up for all the models, and for the Activity Repository to extend that. However, this is out of scope for just now.