Add Calories Burned and Return Property in Model & API
In order for an athlete to see their calories burned from their activity log, this should be calculated according to their weight and speed of their Activity. It should also be returned in the DB Model, and in the API Response when adding an Activity.

This commit ensures the Calories Burned calculation is added, without any magic numbers. This allows the calculation to be modified if necessary, and does not hard-code the values. Returning the `calories_burned` property in the DB model is fine, however, it is an issue when returning the API response. Additionally-set properties are not included when traversing throught the model as a dict. Thus, the work-around (acknowledged as hacky) is to manually create a list of `ADDITIONAL_MODEL_PROPERTIES` to include when serialising the results to JSON in the API.